+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Return over Shopware Commercial plugins is now transfered to Mollie when the Return status is set to "Done" and can be cancelled with the "Cancelled" status. Please note that the refund cannot be cancelled after two hours.
 ### Changes
 - Minimum Supported Shopware version is now 6.4.5.0
+- Add new monolog channel "mollie". You can now add custom handler and assign them to the mollie channel